+++ cloudwanderer/storage_connectors/base_connector.py
@@ -28,3 +28,14 @@ class BaseConnector(ABC):
     @abstractmethod
     def read_resource_of_type_in_account(self, service, resource_type, account_id):
         """Return all resources of this type from this AWS account."""
+
+    @abstractmethod
+    def delete_resource(self, urn):
+        """Delete this resource and all its resource attributes."""
+
+    @abstractmethod
+    def delete_resource_of_type_in_account_region(self, service, resource_type, account_id, region, urns_to_keep=None):
+        """Delete resources of type in account and region unless in list of URNs.
+
+        This is used primarily to clean up old resources.
+        """