+++ tests/integration/cloudwanderer/test_cloud_wanderer_read.py
@@ -2,7 +2,7 @@ import logging
 import unittest
 from unittest.mock import MagicMock
 from cloudwanderer import CloudWanderer, AwsUrn
-from .mocks import generate_mock_session
+from ..mocks import generate_mock_session
 
 
 class TestCloudWandererRead(unittest.TestCase):