+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.5.0
+
+ - Add tests for storage connector
+ - Fixed return of read_all on storage connector
+ - Added delete_resource on storage connector
+ - Added delete_resource_of_type_in_account_region on storage connector
+ - write_resources in CloudWanderer now deletes resources which no longer exist.
+ - Added param for number_of_shards for dynamodb connector rather than hardcoding it.
+
 # 0.4.0
 
 - Introduced more structured `CloudWandererResource` result object to ensure resource attribute keys don't clash with resource keys.